FinalViterbiLogProb completion and test.

+++ src/SmooshableChain.hpp
@@ -44,7 +44,7 @@ class SmooshableChain : public Smooshish {
   const Eigen::MatrixXd& viterbi() const override;
   const Eigen::MatrixXi& viterbi_idx() const override;
   void AuxViterbiPath(int row, int col, std::vector<int>& path) const override;
-  float FinalViterbiLogProb() const;
+  double FinalViterbiLogProb() const;
   std::vector<int> ViterbiPath(int row, int col) const;
   IntVectorVector ViterbiPaths() const;
 };
